WIP_dates_experiences

+++ src/Unetwork/AdminBundle/Entity/Community.php~
@@ -61,6 +61,7 @@ class Community
         $this->actualitys = new ArrayCollection();
         $this->setCreated(new \DateTime());
         $this->setUpdated(new \DateTime());
+        $this->setPath('../../../../../img/community_couverture_default.jpg');
     }
 
     /**
@@ -147,8 +148,12 @@ class Community
      */
     public function removeUpload()
     {
-        if ($file = $this->getAbsolutePath()) {
-            unlink($file);
+        if($this->getPath() != '../../../../../img/community_couverture_default.jpg'){
+
+            if ($file = $this->getAbsolutePath()) {
+                unlink($file);
+            }
+
         }
     }
